Move getCsvTemplateColumns
