+++ src/routes/files.ts
@@ -1,6 +1,8 @@
+import _ from 'lodash';
+
 import { FastifyInstance } from 'fastify';
 import { stringify } from 'csv/sync';
-import { Config } from '../config';
+import { Config, ContactProperty } from '../config';
 import SessionCache from '../services/session-cache';
 import createZip from '../services/files';
 
@@ -8,7 +10,7 @@ export default async function files(fastify: FastifyInstance) {
   fastify.get('/files/template/:placeType', async (req) => {
     const params: any = req.params;
     const placeType = params.placeType;
-    const columns = Config.getCsvTemplateColumns(placeType);
+    const columns = getCsvTemplateColumns(placeType);
     return stringify([columns]);
   });
 
@@ -18,4 +20,22 @@ export default async function files(fastify: FastifyInstance) {
     reply.header('Content-Disposition', `attachment; filename="${Date.now()}_${req.chtSession.authInfo.friendly}_users.zip"`);
     return zip.generateNodeStream();
   });
+}
+
+function getCsvTemplateColumns(placeType: string) {
+  const placeTypeConfig = Config.getContactType(placeType);
+  const hierarchy = Config.getHierarchyWithReplacement(placeTypeConfig);
+  const userRoleConfig = Config.getUserRoleConfig(placeTypeConfig);
+
+  const extractColumns = (properties: ContactProperty[]) => properties
+    .filter(p => p.type !== 'generated')
+    .map(p => p.friendly_name);
+
+  const columns = _.uniq([
+    ...hierarchy.map(p => p.friendly_name),
+    ...extractColumns(placeTypeConfig.place_properties),
+    ...extractColumns(placeTypeConfig.contact_properties),
+    ...(Config.hasMultipleRoles(placeTypeConfig) ? [userRoleConfig.friendly_name] : []),
+  ]);
+  return columns;
 }
